+++ objects/Sphere.cc
@@ -13,7 +13,7 @@
 #include "base/Ray.h"
 #include "scene/Intersect.h"
 
-Sphere::Sphere(const Vec &p, float r, Material m)
+Sphere::Sphere(const Vec &p, double r, Material m)
     : Object(m), position(p), radius(r) {
 }
 
@@ -24,22 +24,22 @@ void Sphere::GetIntersect(const Ray &r, Intersect *out) const {
   bool is_in_sphere = false;
   out->geometry_ptr = NULL;
   Vec v = position.Sub(r.position);
-  float slen_of_v = v.SqrLength();
-  float sqr_radius = radius*radius;
+  double slen_of_v = v.SqrLength();
+  double sqr_radius = radius*radius;
   if (slen_of_v <= sqr_radius) {
     is_in_sphere = true;
   } else if (v.DotProduct(r.direction) <= 0) { // Fixme
     return;
   }
-  float a = r.direction.SqrLength();
-  float b = -2*v.DotProduct(r.direction);
-  float c = a*(slen_of_v - sqr_radius);
-  float delta = b*b - 4*a*c;
+  double a = r.direction.SqrLength();
+  double b = -2*v.DotProduct(r.direction);
+  double c = a*(slen_of_v - sqr_radius);
+  double delta = b*b - 4*a*c;
   if (delta >= 0) {
-    float sqrt_delta = sqrt(delta);
-    float t1 = (-b - sqrt_delta) / (2*a);
-    float t2 = (-b + sqrt_delta) / (2*a);
-    float tmin;
+    double sqrt_delta = sqrt(delta);
+    double t1 = (-b - sqrt_delta) / (2*a);
+    double t2 = (-b + sqrt_delta) / (2*a);
+    double tmin;
     if (t2 <= Object::kThreshold) {
       return;
     } else if (t1 <= Object::kThreshold) {
